refactor(transformer/class-properties): rename file (#7767)
Cosmetic change. Rename file that deals with static property initializers to have a more descriptive name.
