+++ crates/oxc_transformer/src/es2022/class_properties/static_prop_init.rs
